# api/openapi.yaml
openapi: 3.0.3
info:
  title: Enzyme API
  description: REST API for Enzyme, a self-hostable Slack alternative
  version: 1.0.0
servers:
  - url: /api
    description: API server
...
paths:
...
  /workspaces/{wid}/emojis/upload:
    post:
      tags: [emojis]
      summary: Upload a custom emoji
      operationId: uploadCustomEmoji
      security:
        - bearerAuth: []
      parameters:
        - name: wid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, name]
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
      responses:
        '200':
          description: Emoji uploaded
          content:
            application/json:
              schema:
                type: object
                required: [emoji]
                properties:
                  emoji:
                    $ref: '#/components/schemas/CustomEmoji'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /workspaces/{wid}/emojis/list:
    post:
      tags: [emojis]
      summary: List custom emojis for a workspace
      operationId: listCustomEmojis
      security:
        - bearerAuth: []
      parameters:
        - name: wid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of custom emojis
          content:
            application/json:
              schema:
                type: object
                required: [emojis]
                properties:
                  emojis:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomEmoji'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /emojis/{id}/delete:
    post:
      tags: [emojis]
      summary: Delete a custom emoji
      operationId: deleteCustomEmoji
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Emoji deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # User endpoints
  /users/{id}:
    get:
      tags: [users]
      summary: Get user profile
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me/profile:
    post:
      tags: [users]
      summary: Update own profile
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileInput'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
  /users/me/avatar:
    post:
      tags: [users]
      summary: Upload avatar image
      operationId: uploadAvatar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarUploadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: [users]
      summary: Remove avatar
      operationId: deleteAvatar
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Avatar removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
...
    UserProfile:
      type: object
      required: [id, display_name, status, created_at]
      properties:
        id:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    UpdateProfileInput:
      type: object
      properties:
        display_name:
          type: string
    AvatarUploadResponse:
      type: object
      required: [avatar_url]
      properties:
        avatar_url:
          type: string
...
    User:
      type: object
      required: [id, email, display_name, status, created_at, updated_at]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        email_verified_at:
          type: string
          format: date-time
        display_name:
          type: string
        avatar_url:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # Workspace schemas
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ApiErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    CustomEmoji:
      type: object
      required: [id, workspace_id, name, created_by, content_type, size_bytes, url, created_at]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        name:
          type: string
        created_by:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
          format: int64
        url:
          type: string
        created_at:
          type: string
          format: date-time
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...